Replace API Gateway SDK call with DynamoDB lookup to prevent timeout  (#84)
PR is a rework of #81 because that broke development when deployed. In this PR end-to-end tests have been added to validate the new functionality.
What

Use the APIAuthenticatorData DynamoDB table with a new "apiGatewayId" column to get the API Gateway API name from the API Gateway API ID instead of using the AWS ApiGateway SDK.
Remove the now redundant dependencies.
Add proper End-to-End tests for the main flow working with real (Docker) databases without mocking - should catch integration issues like what happened with the last PR.
Why

We identified an issue where under high load requests can time out at the point where we use the API Gateway SDK to fetch API name from API ID - it is suspected that we're hitting a rate limit here and being throttled, and this is not a limit we can increase.

We need to remove the core dependency of using the API Gateway SDK to find the API name from the ID.
How

    Remove Amazon.APIGateway and Amazon.SecurityToken dependencies that were used for getting the API Gateway name from the ID in a given account
    Replace these with the fetched API name from the DynamoDB database in the existing lookup table which already associates names to IDs - use the existing GetTokenData for service auth or the new GetAPIDataByApiIdAsync for user auth.

Important note: After deploying this, the database is the new source of truth for valid API names so we will need to ensure any new APIs that get the updated authorizer are added to the APIAuthenticatorData database table. Previously the API Gateway itself was the source of truth so it would be automatic.